Add portfolio-level production targets

This function calculates the portfolio-level production targets, as calculated using either the technology market share ratio tmsr or sector market share percentage smsp.

add_portfolio_target(data)

Arguments

data
A "data.frame" like the output of summarize_portfolio_production().

Value

A tibble with the same groups as the input (if any) and columns: tmsr_target_weighted_production and smsp_target_weighted_production.

Examples

library(r2dii.analysis)
library(r2dii.data)
library(r2dii.match)

master <- r2dii.data::loanbook_demo %>%
  r2dii.match::match_name(r2dii.data::ald_demo) %>%
  r2dii.match::prioritize() %>%
  join_ald_scenario(r2dii.data::ald_demo,
    r2dii.data::scenario_demo_2020,
    region_isos = region_isos_demo
  )

portfolio_production <- summarize_portfolio_production(master)
portfolio_production

#> # A tibble: 684 x 7
#>    sector     technology  year scenario  tmsr    smsp weighted_production
#>    <chr>      <chr>      <int> <chr>    <dbl>   <dbl>               <dbl>
#>  1 automotive electric    2020 cps       1    0                   148935.
#>  2 automotive electric    2020 sds       1    0                   148935.
#>  3 automotive electric    2020 sps       1    0                   148935.
#>  4 automotive electric    2021 cps       1.12 0.00108             150875.
#>  5 automotive electric    2021 sds       1.16 0.00653             150875.
#>  6 automotive electric    2021 sps       1.14 0.00137             150875.
#>  7 automotive electric    2022 cps       1.24 0.00213             152816.
#>  8 automotive electric    2022 sds       1.32 0.0131              152816.
#>  9 automotive electric    2022 sps       1.29 0.00273             152816.
#> 10 automotive electric    2023 cps       1.35 0.00316             154757.
#> # … with 674 more rows

add_portfolio_target(portfolio_production)

#> # A tibble: 684 x 7
#>    sector technology  year scenario weighted_produc… tmsr_target_wei…
#>    <chr>  <chr>      <int> <chr>               <dbl>            <dbl>
#>  1 autom… electric    2020 cps               148935.          148935.
#>  2 autom… electric    2020 sds               148935.          148935.
#>  3 autom… electric    2020 sps               148935.          148935.
#>  4 autom… electric    2021 cps               150875.          166850.
#>  5 autom… electric    2021 sds               150875.          172982.
#>  6 autom… electric    2021 sps               150875.          170433.
#>  7 autom… electric    2022 cps               152816.          184388.
#>  8 autom… electric    2022 sds               152816.          197029.
#>  9 autom… electric    2022 sps               152816.          191931.
#> 10 autom… electric    2023 cps               154757.          201599.
#> # … with 674 more rows, and 1 more variable:
#> #   smsp_target_weighted_production <dbl>