r2dii.analysis: Measure Climate Scenario Alignment of Corporate Loans

These tools help you to assess if a corporate lending portfolio aligns with climate goals. They summarize key climate indicators attributed to the portfolio (e.g. production, emission factors), and calculate alignment targets based on climate scenarios. They implement in R the last step of the free software 'PACTA' (Paris Agreement Capital Transition Assessment; <https://2degrees-investing.org/>). Financial institutions use 'PACTA' to study how their capital allocation decisions align with climate change mitigation goals.

See also

Author

Maintainer: Jackson Hoffart jackson.hoffart@gmail.com (ORCID)

Authors:

Mauro Lepore maurolepore@gmail.com (ORCID) [contractor]

Klaus Hogedorn klaus@2degrees-investing.org

Other contributors:

2 Degrees Investing Initiative contact@2degrees-investing.org [copyright holder, funder]