Calculate portfolio targets using the Sectoral Decarbonisation Approach (SDA)

The Sectoral Decarbonisation Approach (SDA) is a method for setting corporate emission reduction targets in line with climate science.

sda_portfolio_target(
  market,
  portfolio,
  scenario = "B2DS",
  geography = "Global",
  sector = NULL,
  start_year = NULL,
  target_year = NULL
)

Arguments

market
A dataframe like r2dii.analysis::market.

portfolio
A dataframe like r2dii.analysis::portfolio.

scenario
A character vector giving one or more scenarios to use as the SDA target.

geography
A character vector giving one or more scenario geographies for each scenario.

sector
A character vector giving one or more sectors present in both market and portfolio data. NULL defaults to all expected sectors (see section See Also).

start_year
A length-1 numeric or character vector giving the start year used in the SDA calculation. NULL defaults to extracting the year from a configuration file (see section See Also).

target_year
A length-1 numeric or character vector giving the end year used in the SDA calculation. It must be a year present in all sectors of market. NULL defaults to the latest year shared across all sectors given by sector and found in market.

Value

Returns a dataframe where the scen_sec_emissions_factor column holds the result of the SDA calculation.

See also

Examples

# Use example configuration-files
library(r2dii.utils)

# Use a toy configuration file
restore_options <- options(r2dii_config = example_config("config_demo.yml"))
on.exit(restore_options)

# Use `start_year` from the configuration file
START.YEAR()

#> [1] 2019

sda_portfolio_target(market, portfolio)

#> Warning: Skipping sectors not present in both `market` and `portfolio`:
#> Cement, Power, Oil&Gas, Coal, Aviation, FossilFuels, Shipping, Automotive.

#> * Using `sector`: Steel.

#> * Using `start_year`: 2019.

#> * Using `target_year`: 2040.

#> # A tibble: 24 x 9
#>    Allocation Sector Scenario ScenarioGeograp… Investor.Name Portfolio.Name
#>    <chr>      <chr>  <chr>    <chr>            <chr>         <chr>         
#>  1 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  2 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  3 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  4 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  5 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  6 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  7 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  8 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  9 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#> 10 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#> # … with 14 more rows, and 3 more variables: Scen.Sec.EmissionsFactor <dbl>,
#> #   Year <int>, Plan.Sec.EmissionsFactor <dbl>

sda_portfolio_target(market, portfolio, sector = "Steel")

#> * Using `sector`: Steel.

#> * Using `start_year`: 2019.

#> * Using `target_year`: 2040.

#> # A tibble: 24 x 9
#>    Allocation Sector Scenario ScenarioGeograp… Investor.Name Portfolio.Name
#>    <chr>      <chr>  <chr>    <chr>            <chr>         <chr>         
#>  1 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  2 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  3 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  4 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  5 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  6 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  7 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  8 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  9 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#> 10 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#> # … with 14 more rows, and 3 more variables: Scen.Sec.EmissionsFactor <dbl>,
#> #   Year <int>, Plan.Sec.EmissionsFactor <dbl>

# This configuration file lacks `start_year`
options(r2dii_config = example_config("config-toy.yml"))
START.YEAR()

#> NULL

# Fails
try(sda_portfolio_target(market, portfolio))

#> Warning: Skipping sectors not present in both `market` and `portfolio`:
#> Cement, Power, Oil&Gas, Coal, Aviation, FossilFuels, Shipping, Automotive.

#> * Using `sector`: Steel.

#> Error : `start_year` can't be NULL.
#> Did you forget `start_year` in a configuration file or as an argument?

# Passes
sda_portfolio_target(market, portfolio, start_year = "2019")

#> Warning: Skipping sectors not present in both `market` and `portfolio`:
#> Cement, Power, Oil&Gas, Coal, Aviation, FossilFuels, Shipping, Automotive.

#> * Using `sector`: Steel.

#> * Using `start_year`: 2019.

#> * Using `target_year`: 2040.

#> # A tibble: 24 x 9
#>    Allocation Sector Scenario ScenarioGeograp… Investor.Name Portfolio.Name
#>    <chr>      <chr>  <chr>    <chr>            <chr>         <chr>         
#>  1 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  2 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  3 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  4 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  5 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  6 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  7 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  8 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#>  9 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#> 10 Portfolio… Steel  B2DS     Global           Investor1     Portfolio1    
#> # … with 14 more rows, and 3 more variables: Scen.Sec.EmissionsFactor <dbl>,
#> #   Year <int>, Plan.Sec.EmissionsFactor <dbl>