Summarize the portfolio-weighted production at company level

A simpler wrapper used to calculate the loan-weighted production for each matched company, and aggregate to the company level.

summarize_company_production(data, ..., use_credit_limit = FALSE)

Arguments

data
A "data.frame" like the output of join_ald_scenario().

...
Arguments passed on to summarize_weighted_production

use_credit_limit
Logical vector of length 1. FALSE defaults to using the column loan_size_outstanding. Set to TRUE to use the column loan_size_credit_limit instead.

Value

A tibble with the same groups as the input (if any) and columns: sector, technology, year, and weighted_production.

See also

Examples

library(r2dii.analysis)
library(r2dii.data)
library(r2dii.match)

master <- r2dii.data::loanbook_demo %>%
  r2dii.match::match_name(r2dii.data::ald_demo) %>%
  r2dii.match::prioritize() %>%
  join_ald_scenario(r2dii.data::ald_demo, r2dii.data::scenario_demo_2020)

summarize_company_production(master)

#> # A tibble: 0 x 8
#> # … with 8 variables: sector <chr>, technology <chr>, year <dbl>,
#> #   name_ald <chr>, scenario <chr>, tmsr <dbl>, smsp <dbl>,
#> #   weighted_production <dbl>

summarize_company_production(master, use_credit_limit = TRUE)

#> # A tibble: 0 x 8
#> # … with 8 variables: sector <chr>, technology <chr>, year <dbl>,
#> #   name_ald <chr>, scenario <chr>, tmsr <dbl>, smsp <dbl>,
#> #   weighted_production <dbl>