Summarize production based on the weight of each loan per sector per year

Summarize production based on the weight of each loan per sector per year.

summarize_weighted_production(data, ..., use_credit_limit = FALSE)

Arguments

data
A data frame like the output of join_ald_scenario().

...
Variables to group by.

use_credit_limit
Logical vector of length 1. FALSE defaults to using the column loan_size_outstanding. Set to TRUE to use the column loan_size_credit_limit instead.

Value

A tibble with the same groups as the input (if any) and columns: sector, technology, year, and weighted_production.

See also

Examples

library(r2dii.analysis)
library(r2dii.data)
library(r2dii.match)

master <- r2dii.data::loanbook_demo %>%
  r2dii.match::match_name(r2dii.data::ald_demo) %>%
  r2dii.match::prioritize() %>%
  join_ald_scenario(r2dii.data::ald_demo,
    r2dii.data::scenario_demo_2020,
    region_isos = region_isos_demo
  )

summarize_weighted_production(master)

#> # A tibble: 138 x 4
#>    sector     technology  year weighted_production
#>    <chr>      <chr>      <int>               <dbl>
#>  1 automotive electric    2020             446804.
#>  2 automotive electric    2021             452626.
#>  3 automotive electric    2022             458449.
#>  4 automotive electric    2023             464271.
#>  5 automotive electric    2024             470094.
#>  6 automotive electric    2025             475916.
#>  7 automotive electric    2026             481739.
#>  8 automotive hybrid      2020             425303.
#>  9 automotive hybrid      2021             416061.
#> 10 automotive hybrid      2022             406818.
#> # … with 128 more rows

summarize_weighted_production(master, use_credit_limit = TRUE)

#> # A tibble: 138 x 4
#>    sector     technology  year weighted_production
#>    <chr>      <chr>      <int>               <dbl>
#>  1 automotive electric    2020             422486.
#>  2 automotive electric    2021             426003.
#>  3 automotive electric    2022             429520.
#>  4 automotive electric    2023             433037.
#>  5 automotive electric    2024             436554.
#>  6 automotive electric    2025             440071.
#>  7 automotive electric    2026             443588.
#>  8 automotive hybrid      2020             388781.
#>  9 automotive hybrid      2021             387804.
#> 10 automotive hybrid      2022             386827.
#> # … with 128 more rows