Add targets for production, using the market share approach

This function calculates the portfolio-level production targets, as calculated using the market share approach applied to each relevant climate production forecast.

target_market_share(
  data,
  ald,
  scenario,
  region_isos = r2dii.data::region_isos,
  use_credit_limit = FALSE,
  by_company = FALSE
)

Arguments

data
A "data.frame" like the output of r2dii.match::prioritize().

ald
An asset level data frame like r2dii.data::ald_demo.

scenario
A scenario data frame like r2dii.data::scenario_demo_2020.

region_isos
A data frame like r2dii.data::region_isos (default).

use_credit_limit
Logical vector of length 1. FALSE defaults to using the column loan_size_outstanding. Set to TRUE to use the column loan_size_credit_limit instead.

by_company
Logical vector of length 1. FALSE defaults to outputting weighted_production_value at the portfolio-level. Set to TRUE to output weighted_production_value at the company-level.

Value

A tibble with the summarized columns weighted_production_metric and weighted_production_value. If by_company = TRUE, the output will also have the column name_ald.

Handling grouped data

This function ignores existing groups and outputs ungrouped data.

See also

Examples

library(r2dii.analysis)
library(r2dii.data)
library(r2dii.match)

match_result <- r2dii.data::loanbook_demo %>%
  r2dii.match::match_name(r2dii.data::ald_demo) %>%
  r2dii.match::prioritize()

# calculate targets at portfolio level
target_market_share(match_result,
  ald = r2dii.data::ald_demo,
  scenario = r2dii.data::scenario_demo_2020,
  region_isos = r2dii.data::region_isos_demo
)

#> # A tibble: 1,170 x 7
#>    sector technology  year region scenario_source weighted_produc…
#>    <chr>  <chr>      <int> <chr>  <chr>           <chr>           
#>  1 autom… electric    2020 global demo_2020       projected       
#>  2 autom… electric    2020 global demo_2020       normalized_corp…
#>  3 autom… electric    2020 global demo_2020       target_cps      
#>  4 autom… electric    2020 global demo_2020       target_sds      
#>  5 autom… electric    2020 global demo_2020       target_sps      
#>  6 autom… hybrid      2020 global demo_2020       projected       
#>  7 autom… hybrid      2020 global demo_2020       normalized_corp…
#>  8 autom… hybrid      2020 global demo_2020       target_cps      
#>  9 autom… hybrid      2020 global demo_2020       target_sds      
#> 10 autom… hybrid      2020 global demo_2020       target_sps      
#> # … with 1,160 more rows, and 1 more variable: weighted_production_value <dbl>

# calculate targets at company level
target_market_share(match_result,
  ald = r2dii.data::ald_demo,
  scenario = r2dii.data::scenario_demo_2020,
  region_isos = r2dii.data::region_isos_demo,
  by_company = TRUE
)

#> # A tibble: 15,945 x 8
#>    sector technology  year region scenario_source name_ald weighted_produc…
#>    <chr>  <chr>      <int> <chr>  <chr>           <chr>    <chr>           
#>  1 autom… electric    2020 global demo_2020       shangha… projected       
#>  2 autom… electric    2020 global demo_2020       shangha… normalized_corp…
#>  3 autom… electric    2020 global demo_2020       shangha… target_cps      
#>  4 autom… electric    2020 global demo_2020       shangha… target_sds      
#>  5 autom… electric    2020 global demo_2020       shangha… target_sps      
#>  6 autom… electric    2020 global demo_2020       sichuan… projected       
#>  7 autom… electric    2020 global demo_2020       sichuan… normalized_corp…
#>  8 autom… electric    2020 global demo_2020       sichuan… target_cps      
#>  9 autom… electric    2020 global demo_2020       sichuan… target_sds      
#> 10 autom… electric    2020 global demo_2020       sichuan… target_sps      
#> # … with 15,935 more rows, and 1 more variable: weighted_production_value <dbl>