Add targets for CO2 emissions per unit production at the portfolio level, using the SDA approach

This function calculates targets of CO2 emissions per unit production at the portfolio-level, otherwise referred to as "emissions factors". It uses the sectoral-decarbonization approach (SDA) to calculate these targets.

target_sda(
  data,
  ald,
  co2_intensity_scenario,
  use_credit_limit = FALSE,
  by_company = FALSE
)

Arguments

data
A dataframe like the output of r2dii.match::prioritize().

ald
An asset-level data frame like r2dii.data::ald_demo.

co2_intensity_scenario
A scenario data frame like r2dii.data::co2_intensity_scenario_demo.

use_credit_limit
Logical vector of length 1. FALSE defaults to using the column loan_size_outstanding. Set to TRUE to instead use the column loan_size_credit_limit.

by_company
Logical vector of length 1. FALSE defaults to outputting weighted_production_value at the portfolio-level. Set to TRUE to output weighted_production_value at the company-level.

Value

A tibble with the CO2 emissions factors attributed to the portfolio. These values include the portfolio's actual projected CO2 emissions factors, the scenario pathway CO2 emissions factors and the SDA calculated portfolio target emissions factors (see column emission_factor_metric).

Handling grouped data

This function ignores existing groups and outputs ungrouped data.

See also

Examples

installed <- requireNamespace("r2dii.match", quietly = TRUE) &&
  requireNamespace("r2dii.data", quietly = TRUE)
if (!installed) stop("Please install r2dii.match and r2dii.data")

library(r2dii.match)
library(r2dii.data)

# Example datasets from r2dii.data
loanbook <- head(loanbook_demo, 150)
ald <- head(ald_demo, 100)

co2_scenario <- co2_intensity_scenario_demo

# WARNING: Remember to validate matches (see `?prioritize`)
matched <- prioritize(match_name(loanbook, ald))

# You may need to clean your data
anyNA(ald$emission_factor)

#> [1] TRUE

try(target_sda(matched, ald, co2_intensity_scenario = co2_scenario))

#> Warning: Removing ald rows where `emission_factor` is NA

#> # A tibble: 171 x 4
#>    sector    year emission_factor_metric emission_factor_value
#>    <chr>    <dbl> <chr>                                  <dbl>
#>  1 shipping  2013 projected                          39034906.
#>  2 shipping  2014 projected                          38199125.
#>  3 shipping  2015 projected                          37363343.
#>  4 shipping  2016 projected                          36527562.
#>  5 shipping  2017 projected                          35691781.
#>  6 shipping  2018 projected                          34856000.
#>  7 shipping  2019 projected                          34020218.
#>  8 shipping  2020 projected                          33184437.
#>  9 shipping  2021 projected                          32348656.
#> 10 shipping  2022 projected                          31512874.
#> # … with 161 more rows

ald2 <- subset(ald, !is.na(emission_factor))
anyNA(ald2$emission_factor)

#> [1] FALSE

out <- target_sda(matched, ald2, co2_intensity_scenario = co2_scenario)

# The output includes the portfolio's actual projected emissions factors, the
# scenario pathway emissions factors, and the portfolio's target emissions
# factors.
out

#> # A tibble: 171 x 4
#>    sector    year emission_factor_metric emission_factor_value
#>    <chr>    <dbl> <chr>                                  <dbl>
#>  1 shipping  2013 projected                          39034906.
#>  2 shipping  2014 projected                          38199125.
#>  3 shipping  2015 projected                          37363343.
#>  4 shipping  2016 projected                          36527562.
#>  5 shipping  2017 projected                          35691781.
#>  6 shipping  2018 projected                          34856000.
#>  7 shipping  2019 projected                          34020218.
#>  8 shipping  2020 projected                          33184437.
#>  9 shipping  2021 projected                          32348656.
#> 10 shipping  2022 projected                          31512874.
#> # … with 161 more rows

# Split-view by metric
split(out, out$emission_factor_metric)

#> $adjusted_scenario_demo
#> # A tibble: 31 x 4
#>    sector  year emission_factor_metric emission_factor_value
#>    <chr>  <dbl> <chr>                                  <dbl>
#>  1 cement  2020 adjusted_scenario_demo                 0.643
#>  2 cement  2021 adjusted_scenario_demo                 0.588
#>  3 cement  2022 adjusted_scenario_demo                 0.532
#>  4 cement  2023 adjusted_scenario_demo                 0.477
#>  5 cement  2024 adjusted_scenario_demo                 0.422
#>  6 cement  2025 adjusted_scenario_demo                 0.367
#>  7 cement  2026 adjusted_scenario_demo                 0.365
#>  8 cement  2027 adjusted_scenario_demo                 0.364
#>  9 cement  2028 adjusted_scenario_demo                 0.362
#> 10 cement  2029 adjusted_scenario_demo                 0.360
#> # … with 21 more rows
#> 
#> $corporate_economy
#> # A tibble: 39 x 4
#>    sector  year emission_factor_metric emission_factor_value
#>    <chr>  <dbl> <chr>                                  <dbl>
#>  1 cement  2013 corporate_economy                      0.643
#>  2 cement  2014 corporate_economy                      0.654
#>  3 cement  2015 corporate_economy                      0.666
#>  4 cement  2016 corporate_economy                      0.677
#>  5 cement  2017 corporate_economy                      0.689
#>  6 cement  2018 corporate_economy                      0.700
#>  7 cement  2019 corporate_economy                      0.712
#>  8 cement  2020 corporate_economy                      0.723
#>  9 cement  2021 corporate_economy                      0.735
#> 10 cement  2022 corporate_economy                      0.746
#> # … with 29 more rows
#> 
#> $projected
#> # A tibble: 13 x 4
#>    sector    year emission_factor_metric emission_factor_value
#>    <chr>    <dbl> <chr>                                  <dbl>
#>  1 shipping  2013 projected                          39034906.
#>  2 shipping  2014 projected                          38199125.
#>  3 shipping  2015 projected                          37363343.
#>  4 shipping  2016 projected                          36527562.
#>  5 shipping  2017 projected                          35691781.
#>  6 shipping  2018 projected                          34856000.
#>  7 shipping  2019 projected                          34020218.
#>  8 shipping  2020 projected                          33184437.
#>  9 shipping  2021 projected                          32348656.
#> 10 shipping  2022 projected                          31512874.
#> 11 shipping  2023 projected                          30677093.
#> 12 shipping  2024 projected                          29841312.
#> 13 shipping  2025 projected                          29005531.
#> 
#> $target_demo
#> # A tibble: 44 x 4
#>    sector    year emission_factor_metric emission_factor_value
#>    <chr>    <dbl> <chr>                                  <dbl>
#>  1 shipping  2013 target_demo                               NA
#>  2 shipping  2014 target_demo                               NA
#>  3 shipping  2015 target_demo                               NA
#>  4 shipping  2016 target_demo                               NA
#>  5 shipping  2017 target_demo                               NA
#>  6 shipping  2018 target_demo                               NA
#>  7 shipping  2019 target_demo                               NA
#>  8 shipping  2020 target_demo                               NA
#>  9 cement    2020 target_demo                               NA
#> 10 shipping  2021 target_demo                               NA
#> # … with 34 more rows
#> 
#> $target_NA
#> # A tibble: 44 x 4
#>    sector    year emission_factor_metric emission_factor_value
#>    <chr>    <dbl> <chr>                                  <dbl>
#>  1 shipping  2013 target_NA                                 NA
#>  2 shipping  2014 target_NA                                 NA
#>  3 shipping  2015 target_NA                                 NA
#>  4 shipping  2016 target_NA                                 NA
#>  5 shipping  2017 target_NA                                 NA
#>  6 shipping  2018 target_NA                                 NA
#>  7 shipping  2019 target_NA                                 NA
#>  8 shipping  2020 target_NA                                 NA
#>  9 cement    2020 target_NA                                 NA
#> 10 shipping  2021 target_NA                                 NA
#> # … with 34 more rows
#> 

# Calculate company-level targets
out <- target_sda(
  matched, ald2,
  co2_intensity_scenario = co2_scenario,
  by_company = TRUE
)
out

#> # A tibble: 171 x 5
#>    sector    year name_ald            emission_factor_metr… emission_factor_val…
#>    <chr>    <dbl> <chr>               <chr>                                <dbl>
#>  1 shipping  2013 africon shipping sa projected                        39034906.
#>  2 shipping  2014 africon shipping sa projected                        38199125.
#>  3 shipping  2015 africon shipping sa projected                        37363343.
#>  4 shipping  2016 africon shipping sa projected                        36527562.
#>  5 shipping  2017 africon shipping sa projected                        35691781.
#>  6 shipping  2018 africon shipping sa projected                        34856000.
#>  7 shipping  2019 africon shipping sa projected                        34020218.
#>  8 shipping  2020 africon shipping sa projected                        33184437.
#>  9 shipping  2021 africon shipping sa projected                        32348656.
#> 10 shipping  2022 africon shipping sa projected                        31512874.
#> # … with 161 more rows