Join a data-loanbook object to the abcd and scenario

join_abcd_scenario() is a simple wrapper of several calls to dplyr::join_*(), forming the master dataset to be used in later steps of the analysis.

Usage

join_abcd_scenario(
  data,
  abcd,
  scenario,
  region_isos = r2dii.data::region_isos,
  add_green_technologies = FALSE
)

Arguments

data

A data frame like the output of r2dii.match::prioritize.

abcd

An asset level data frame like r2dii.data::abcd_demo.

scenario

A scenario data frame like r2dii.data::scenario_demo_2020.

region_isos

A data frame like r2dii.data::region_isos (default).

add_green_technologies

Logical vector of length 1. FALSE defaults to outputting only technologies that are present in both data and abcd. Set to FALSE to add rows of all possible green technologies (with 0 production).

Value

Returns a fully joined data frame, linking portfolio, abcd and scenario.

See also

Examples

installed <- requireNamespace("r2dii.data", quietly = TRUE) &&
  requireNamespace("r2dii.match", versionCheck = "0.1.0", quietly = TRUE) &&
  packageVersion("r2dii.match") >= "0.1.0"

if (installed) {
  library(r2dii.data)
  library(r2dii.match)

  valid_matches <- match_name(loanbook_demo, abcd_demo) %>%
    # WARNING: Remember to validate matches (see `?prioritize`)
    prioritize()

  valid_matches %>%
    join_abcd_scenario(
      abcd = abcd_demo,
      scenario = scenario_demo_2020,
      region_isos = region_isos_demo
    )
}
#> # A tibble: 9,099 × 45
#>    id_loan id_direct_loantaker name_direct_loantaker  id_intermediate_parent_1
#>    <chr>   <chr>               <chr>                  <chr>                   
#>  1 L6      C304                Yukon Development Corp NA                      
#>  2 L6      C304                Yukon Development Corp NA                      
#>  3 L6      C304                Yukon Development Corp NA                      
#>  4 L6      C304                Yukon Development Corp NA                      
#>  5 L6      C304                Yukon Development Corp NA                      
#>  6 L6      C304                Yukon Development Corp NA                      
#>  7 L6      C304                Yukon Development Corp NA                      
#>  8 L6      C304                Yukon Development Corp NA                      
#>  9 L6      C304                Yukon Development Corp NA                      
#> 10 L6      C304                Yukon Development Corp NA                      
#> # ℹ 9,089 more rows
#> # ℹ 41 more variables: name_intermediate_parent_1 <chr>,
#> #   id_ultimate_parent <chr>, name_ultimate_parent <chr>,
#> #   loan_size_outstanding <dbl>, loan_size_outstanding_currency <chr>,
#> #   loan_size_credit_limit <dbl>, loan_size_credit_limit_currency <chr>,
#> #   sector_classification_system <chr>, sector_classification_input_type <chr>,
#> #   sector_classification_direct_loantaker <dbl>, fi_type <chr>, …