Summaries based on the weight of each loan per sector per year

Based on on the weight of each loan per sector per year, summarize_weighted_production() and summarize_weighted_percent_change() summarize the production and percent-change, respectively.

summarize_weighted_production(data, ..., use_credit_limit = FALSE)

summarize_weighted_percent_change(data, ..., use_credit_limit = FALSE)

Arguments

data
A data frame like the output of join_ald_scenario().

...
Variables to group by.

use_credit_limit
Logical vector of length 1. FALSE defaults to using the column loan_size_outstanding. Set to TRUE to instead use the column loan_size_credit_limit.

Value

A tibble with the same groups as the input (if any) and columns: sector, technology, and year; and weighted_production or weighted_production for summarize_weighted_production() and summarize_weighted_percent_change(), respectively.

Warning

The percent-change analysis excludes companies with 0 production. percent-change is undefined for companies that have no initial production; including such companies would cause percent-change percentage to be infinite, which is wrong.

See also

Examples

installed <- requireNamespace("r2dii.data", quietly = TRUE) &&
  requireNamespace("r2dii.match", quietly = TRUE)
if (!installed) stop("Please install r2dii.match and r2dii.data")

library(r2dii.data)
library(r2dii.match)

master <- loanbook_demo %>%
  match_name(ald_demo) %>%
  prioritize() %>%
  join_ald_scenario(
    ald = ald_demo,
    scenario = scenario_demo_2020,
    region_isos = region_isos_demo
  )

summarize_weighted_production(master)

#> # A tibble: 138 x 5
#>    sector_ald technology  year weighted_production weighted_technology_share
#>    <chr>      <chr>      <int>               <dbl>                     <dbl>
#>  1 automotive electric    2020             973775.                     0.114
#>  2 automotive electric    2021            1018967.                     0.118
#>  3 automotive electric    2022            1064159.                     0.122
#>  4 automotive electric    2023            1109351.                     0.126
#>  5 automotive electric    2024            1154543.                     0.130
#>  6 automotive electric    2025            1199735.                     0.133
#>  7 automotive electric    2026            1244926.                     0.137
#>  8 automotive hybrid      2020            1886042.                     0.221
#>  9 automotive hybrid      2021            1709634.                     0.198
#> 10 automotive hybrid      2022            1533226.                     0.176
#> # … with 128 more rows

summarize_weighted_production(master, use_credit_limit = TRUE)

#> # A tibble: 138 x 5
#>    sector_ald technology  year weighted_production weighted_technology_share
#>    <chr>      <chr>      <int>               <dbl>                     <dbl>
#>  1 automotive electric    2020             973775.                     0.114
#>  2 automotive electric    2021            1018967.                     0.118
#>  3 automotive electric    2022            1064159.                     0.122
#>  4 automotive electric    2023            1109351.                     0.126
#>  5 automotive electric    2024            1154543.                     0.130
#>  6 automotive electric    2025            1199735.                     0.133
#>  7 automotive electric    2026            1244926.                     0.137
#>  8 automotive hybrid      2020            1886042.                     0.221
#>  9 automotive hybrid      2021            1709634.                     0.198
#> 10 automotive hybrid      2022            1533226.                     0.176
#> # … with 128 more rows

summarize_weighted_percent_change(master)

#> # A tibble: 138 x 4
#>    sector_ald technology  year weighted_percent_change
#>    <chr>      <chr>      <int>                   <dbl>
#>  1 automotive electric    2020                  0     
#>  2 automotive electric    2021                  0.0881
#>  3 automotive electric    2022                  0.176 
#>  4 automotive electric    2023                  0.264 
#>  5 automotive electric    2024                  0.352 
#>  6 automotive electric    2025                  0.440 
#>  7 automotive electric    2026                  0.528 
#>  8 automotive hybrid      2020                  1.78  
#>  9 automotive hybrid      2021                  1.43  
#> 10 automotive hybrid      2022                  1.09  
#> # … with 128 more rows

summarize_weighted_percent_change(master, use_credit_limit = TRUE)

#> # A tibble: 138 x 4
#>    sector_ald technology  year weighted_percent_change
#>    <chr>      <chr>      <int>                   <dbl>
#>  1 automotive electric    2020                  0     
#>  2 automotive electric    2021                  0.0881
#>  3 automotive electric    2022                  0.176 
#>  4 automotive electric    2023                  0.264 
#>  5 automotive electric    2024                  0.352 
#>  6 automotive electric    2025                  0.440 
#>  7 automotive electric    2026                  0.528 
#>  8 automotive hybrid      2020                  1.78  
#>  9 automotive hybrid      2021                  1.43  
#> 10 automotive hybrid      2022                  1.09  
#> # … with 128 more rows