Add targets for production, using the market share approach

This function calculates the portfolio-level production targets, as calculated using the market share approach applied to each relevant climate production forecast.

target_market_share(
  data,
  ald,
  scenario,
  region_isos = r2dii.data::region_isos,
  use_credit_limit = FALSE,
  by_company = FALSE,
  weight_production = TRUE
)

Arguments

data
A "data.frame" like the output of r2dii.match::prioritize().

ald
An asset level data frame like r2dii.data::ald_demo.

scenario
A scenario data frame like r2dii.data::scenario_demo_2020.

region_isos
A data frame like r2dii.data::region_isos (default).

use_credit_limit
Logical vector of length 1. FALSE defaults to using the column loan_size_outstanding. Set to TRUE to use the column loan_size_credit_limit instead.

by_company
Logical vector of length 1. FALSE defaults to outputting production_value at the portfolio-level. Set to TRUE to output production_value at the company-level.

weight_production
Logical vector of length 1. TRUE defaults to outputting production, weighted by relative loan-size. Set to FALSE to output the unweighted production values.

Value

A tibble with the summarized columns metric, production and technology_share. If by_company = TRUE, the output will also have the column name_ald.

Handling grouped data

This function ignores existing groups and outputs ungrouped data.

See also

Examples

installed <- requireNamespace("r2dii.data", quietly = TRUE) &&
  requireNamespace("r2dii.match", quietly = TRUE)
if (!installed) stop("Please install r2dii.match and r2dii.data")

library(r2dii.data)
library(r2dii.match)

matched <- loanbook_demo %>%
  match_name(ald_demo) %>%
  prioritize()

# Calculate targets at portfolio level
matched %>%
  target_market_share(
    ald = ald_demo,
    scenario = scenario_demo_2020,
    region_isos = region_isos_demo
  )

#> # A tibble: 3,642 x 8
#>    sector     technology  year region scenario_source metric     production
#>    <chr>      <chr>      <int> <chr>  <chr>           <chr>           <dbl>
#>  1 automotive electric    2020 global demo_2020       projected     324592.
#>  2 automotive electric    2020 global demo_2020       target_cps    324592.
#>  3 automotive electric    2020 global demo_2020       target_sds    324592.
#>  4 automotive electric    2020 global demo_2020       target_sps    324592.
#>  5 automotive electric    2021 global demo_2020       projected     339656.
#>  6 automotive electric    2021 global demo_2020       target_cps    329191.
#>  7 automotive electric    2021 global demo_2020       target_sds    352505.
#>  8 automotive electric    2021 global demo_2020       target_sps    330435.
#>  9 automotive electric    2022 global demo_2020       projected     354720.
#> 10 automotive electric    2022 global demo_2020       target_cps    333693.
#> # … with 3,632 more rows, and 1 more variable: technology_share <dbl>

# Calculate targets at company level
matched %>%
  target_market_share(
    ald = ald_demo,
    scenario = scenario_demo_2020,
    region_isos = region_isos_demo,
    by_company = TRUE
  )

#> Warning: You've supplied `by_company = TRUE` and `weight_production = TRUE`.
#> This will result in company-level results, weighted by the portfolio
#> loan size, which is rarely useful. Did you mean to set one of these
#> arguments to `FALSE`?

#> # A tibble: 14,754 x 11
#>    sector  technology  year region scenario_source name_ald   sector_weighted_p…
#>    <chr>   <chr>      <int> <chr>  <chr>           <chr>                   <dbl>
#>  1 automo… electric    2020 global demo_2020       toyota mo…           4275858.
#>  2 automo… electric    2020 global demo_2020       toyota mo…           4275858.
#>  3 automo… electric    2020 global demo_2020       toyota mo…           4275858.
#>  4 automo… electric    2020 global demo_2020       toyota mo…           4275858.
#>  5 automo… hybrid      2020 global demo_2020       toyota mo…           4275858.
#>  6 automo… hybrid      2020 global demo_2020       toyota mo…           4275858.
#>  7 automo… hybrid      2020 global demo_2020       toyota mo…           4275858.
#>  8 automo… hybrid      2020 global demo_2020       toyota mo…           4275858.
#>  9 automo… ice         2020 global demo_2020       toyota mo…           4275858.
#> 10 automo… ice         2020 global demo_2020       toyota mo…           4275858.
#> # … with 14,744 more rows, and 4 more variables:
#> #   technology_weighted_production <dbl>, metric <chr>, production <dbl>,
#> #   technology_share <dbl>

matched %>%
  target_market_share(
    ald = ald_demo,
    scenario = scenario_demo_2020,
    region_isos = region_isos_demo,
    # Calculate unweighted targets
    weight_production = FALSE
  )

#> # A tibble: 3,642 x 8
#>    sector     technology  year region scenario_source metric     production
#>    <chr>      <chr>      <int> <chr>  <chr>           <chr>           <dbl>
#>  1 automotive electric    2020 global demo_2020       projected     324592.
#>  2 automotive electric    2020 global demo_2020       target_cps    324592.
#>  3 automotive electric    2020 global demo_2020       target_sds    324592.
#>  4 automotive electric    2020 global demo_2020       target_sps    324592.
#>  5 automotive electric    2021 global demo_2020       projected     339656.
#>  6 automotive electric    2021 global demo_2020       target_cps    329191.
#>  7 automotive electric    2021 global demo_2020       target_sds    352505.
#>  8 automotive electric    2021 global demo_2020       target_sps    330435.
#>  9 automotive electric    2022 global demo_2020       projected     354720.
#> 10 automotive electric    2022 global demo_2020       target_cps    333693.
#> # … with 3,632 more rows, and 1 more variable: technology_share <dbl>