Add targets for CO2 emissions per unit production at the portfolio level, using the SDA approach

This function calculates targets of CO2 emissions per unit production at the portfolio-level, otherwise referred to as "emissions factors". It uses the sectoral-decarbonization approach (SDA) to calculate these targets.

target_sda(data, ald, co2_intensity_scenario, use_credit_limit = FALSE)

Arguments

data
A dataframe like the output of r2dii.match::prioritize().

ald
An asset-level data frame like r2dii.data::ald_demo.

co2_intensity_scenario
A scenario data frame like r2dii.data::co2_intensity_scenario_demo.

use_credit_limit
Logical vector of length 1. FALSE defaults to using the column loan_size_outstanding. Set to TRUE to instead use the column loan_size_credit_limit.

Value

A tibble with the CO2 emissions factors attributed to the portfolio. These values include the portfolio's actual projected CO2 emissions factors, the scenario pathway CO2 emissions factors and the SDA calculated portfolio target emissions factors (see column emission_factor_metric).

Handling grouped data

This function ignores existing groups and outputs ungrouped data.

See also

Examples

installed <- requireNamespace("r2dii.data", quietly = TRUE) &&
  requireNamespace("r2dii.match", quietly = TRUE)
if (!installed) stop("Please install r2dii.match and r2dii.data")

library(r2dii.data)
library(r2dii.match)

valid_matches <- match_name(loanbook_demo, ald_demo) %>%
  # WARNING: Remember to validate matches (see `?prioritize`)
  prioritize()

out <- valid_matches %>%
  target_sda(
    ald = ald_demo,
    co2_intensity_scenario = co2_intensity_scenario_demo
  )

# The output includes the portfolio's actual projected emissions factors, the
# scenario pathway emissions factors, and the portfolio's target emissions
# factors.
out

#> # A tibble: 28 x 4
#>    sector  year emission_factor_metric emission_factor_value
#>    <chr>  <dbl> <chr>                                  <dbl>
#>  1 cement  2020 projected                              0.664
#>  2 cement  2020 target                                 0.669
#>  3 cement  2020 scenario_benchmark                     0.7  
#>  4 cement  2021 projected                              0.665
#>  5 cement  2021 target                                 0.612
#>  6 cement  2021 scenario_benchmark                     0.64 
#>  7 cement  2022 projected                              0.666
#>  8 cement  2022 target                                 0.555
#>  9 cement  2022 scenario_benchmark                     0.580
#> 10 cement  2023 projected                              0.667
#> # … with 18 more rows

# Split view by metric
split(out, out$emission_factor_metric)

#> $projected
#> # A tibble: 6 x 4
#>   sector  year emission_factor_metric emission_factor_value
#>   <chr>  <dbl> <chr>                                  <dbl>
#> 1 cement  2020 projected                              0.664
#> 2 cement  2021 projected                              0.665
#> 3 cement  2022 projected                              0.666
#> 4 cement  2023 projected                              0.667
#> 5 cement  2024 projected                              0.668
#> 6 cement  2025 projected                              0.669
#> 
#> $scenario_benchmark
#> # A tibble: 11 x 4
#>    sector  year emission_factor_metric emission_factor_value
#>    <chr>  <dbl> <chr>                                  <dbl>
#>  1 cement  2020 scenario_benchmark                     0.7  
#>  2 cement  2021 scenario_benchmark                     0.64 
#>  3 cement  2022 scenario_benchmark                     0.580
#>  4 cement  2023 scenario_benchmark                     0.52 
#>  5 cement  2024 scenario_benchmark                     0.46 
#>  6 cement  2025 scenario_benchmark                     0.4  
#>  7 cement  2030 scenario_benchmark                     0.39 
#>  8 cement  2035 scenario_benchmark                     0.38 
#>  9 cement  2040 scenario_benchmark                     0.25 
#> 10 cement  2045 scenario_benchmark                     0.2  
#> 11 cement  2050 scenario_benchmark                     0.1  
#> 
#> $target
#> # A tibble: 11 x 4
#>    sector  year emission_factor_metric emission_factor_value
#>    <chr>  <dbl> <chr>                                  <dbl>
#>  1 cement  2020 target                                 0.669
#>  2 cement  2021 target                                 0.612
#>  3 cement  2022 target                                 0.555
#>  4 cement  2023 target                                 0.498
#>  5 cement  2024 target                                 0.441
#>  6 cement  2025 target                                 0.384
#>  7 cement  2030 target                                 0.375
#>  8 cement  2035 target                                 0.365
#>  9 cement  2040 target                                 0.242
#> 10 cement  2045 target                                 0.195
#> 11 cement  2050 target                                 0.1  
#>