Add targets for CO2 emissions per unit production at the portfolio level, using the SDA approach

This function calculates targets of CO2 emissions per unit production at the portfolio-level, otherwise referred to as "emissions factors". It uses the sectoral-decarbonization approach (SDA) to calculate these targets.

target_sda(
  data,
  ald,
  co2_intensity_scenario,
  use_credit_limit = FALSE,
  by_company = FALSE
)

Arguments

data
A dataframe like the output of r2dii.match::prioritize().

ald
An asset-level data frame like r2dii.data::ald_demo.

co2_intensity_scenario
A scenario data frame like r2dii.data::co2_intensity_scenario_demo.

use_credit_limit
Logical vector of length 1. FALSE defaults to using the column loan_size_outstanding. Set to TRUE to instead use the column loan_size_credit_limit.

by_company
Logical vector of length 1. FALSE defaults to outputting weighted_production_value at the portfolio-level. Set to TRUE to output weighted_production_value at the company-level.

Value

A tibble with the CO2 emissions factors attributed to the portfolio. These values include the portfolio's actual projected CO2 emissions factors, the scenario pathway CO2 emissions factors and the SDA calculated portfolio target emissions factors (see column emission_factor_metric).

Handling grouped data

This function ignores existing groups and outputs ungrouped data.

See also

Examples

installed <- requireNamespace("r2dii.match", quietly = TRUE) &&
  requireNamespace("r2dii.data", quietly = TRUE)
if (!installed) stop("Please install r2dii.match and r2dii.data")

library(r2dii.match)
library(r2dii.data)

# Example datasets from r2dii.data
loanbook <- loanbook_demo
ald <- ald_demo
co2_scenario <- co2_intensity_scenario_demo

# WARNING: Remember to validate matches (see `?prioritize`)
matched <- prioritize(match_name(loanbook, ald))

# You may need to clean your data
anyNA(ald$emission_factor)

#> [1] TRUE

try(target_sda(matched, ald, co2_intensity_scenario = co2_scenario))

#> Warning: Removing ald rows where `emission_factor` is NA

#> # A tibble: 508 x 4
#>    sector      year emission_factor_metric emission_factor_value
#>    <chr>      <dbl> <chr>                                  <dbl>
#>  1 automotive  2002 projected                              0.397
#>  2 automotive  2003 projected                              0.389
#>  3 automotive  2004 projected                              0.381
#>  4 automotive  2005 projected                              0.373
#>  5 automotive  2006 projected                              0.365
#>  6 automotive  2007 projected                              0.358
#>  7 automotive  2008 projected                              0.350
#>  8 automotive  2009 projected                              0.342
#>  9 automotive  2010 projected                              0.334
#> 10 automotive  2011 projected                              0.326
#> # … with 498 more rows

ald2 <- subset(ald, !is.na(emission_factor))
anyNA(ald2$emission_factor)

#> [1] FALSE

out <- target_sda(matched, ald2, co2_intensity_scenario = co2_scenario)

# The output includes the portfolio's actual projected emissions factors, the
# scenario pathway emissions factors, and the portfolio's target emissions
# factors.
out

#> # A tibble: 508 x 4
#>    sector      year emission_factor_metric emission_factor_value
#>    <chr>      <dbl> <chr>                                  <dbl>
#>  1 automotive  2002 projected                              0.397
#>  2 automotive  2003 projected                              0.389
#>  3 automotive  2004 projected                              0.381
#>  4 automotive  2005 projected                              0.373
#>  5 automotive  2006 projected                              0.365
#>  6 automotive  2007 projected                              0.358
#>  7 automotive  2008 projected                              0.350
#>  8 automotive  2009 projected                              0.342
#>  9 automotive  2010 projected                              0.334
#> 10 automotive  2011 projected                              0.326
#> # … with 498 more rows

# Split-view by metric
split(out, out$emission_factor_metric)

#> $adjusted_scenario_demo
#> # A tibble: 62 x 4
#>    sector  year emission_factor_metric emission_factor_value
#>    <chr>  <dbl> <chr>                                  <dbl>
#>  1 cement  2020 adjusted_scenario_demo                 0.661
#>  2 steel   2020 adjusted_scenario_demo                 0.262
#>  3 cement  2021 adjusted_scenario_demo                 0.605
#>  4 steel   2021 adjusted_scenario_demo                 0.249
#>  5 cement  2022 adjusted_scenario_demo                 0.548
#>  6 steel   2022 adjusted_scenario_demo                 0.236
#>  7 cement  2023 adjusted_scenario_demo                 0.491
#>  8 steel   2023 adjusted_scenario_demo                 0.222
#>  9 cement  2024 adjusted_scenario_demo                 0.435
#> 10 steel   2024 adjusted_scenario_demo                 0.209
#> # … with 52 more rows
#> 
#> $corporate_economy
#> # A tibble: 103 x 4
#>    sector      year emission_factor_metric emission_factor_value
#>    <chr>      <dbl> <chr>                                  <dbl>
#>  1 automotive  2002 corporate_economy                      0.239
#>  2 automotive  2003 corporate_economy                      0.240
#>  3 automotive  2004 corporate_economy                      0.241
#>  4 automotive  2005 corporate_economy                      0.242
#>  5 automotive  2006 corporate_economy                      0.244
#>  6 automotive  2007 corporate_economy                      0.245
#>  7 automotive  2008 corporate_economy                      0.246
#>  8 automotive  2009 corporate_economy                      0.248
#>  9 automotive  2010 corporate_economy                      0.249
#> 10 automotive  2011 corporate_economy                      0.251
#> # … with 93 more rows
#> 
#> $projected
#> # A tibble: 77 x 4
#>    sector      year emission_factor_metric emission_factor_value
#>    <chr>      <dbl> <chr>                                  <dbl>
#>  1 automotive  2002 projected                              0.397
#>  2 automotive  2003 projected                              0.389
#>  3 automotive  2004 projected                              0.381
#>  4 automotive  2005 projected                              0.373
#>  5 automotive  2006 projected                              0.365
#>  6 automotive  2007 projected                              0.358
#>  7 automotive  2008 projected                              0.350
#>  8 automotive  2009 projected                              0.342
#>  9 automotive  2010 projected                              0.334
#> 10 automotive  2011 projected                              0.326
#> # … with 67 more rows
#> 
#> $target_demo
#> # A tibble: 133 x 4
#>    sector      year emission_factor_metric emission_factor_value
#>    <chr>      <dbl> <chr>                                  <dbl>
#>  1 automotive  2002 target_demo                               NA
#>  2 automotive  2003 target_demo                               NA
#>  3 automotive  2004 target_demo                               NA
#>  4 automotive  2005 target_demo                               NA
#>  5 automotive  2006 target_demo                               NA
#>  6 automotive  2007 target_demo                               NA
#>  7 automotive  2008 target_demo                               NA
#>  8 automotive  2009 target_demo                               NA
#>  9 automotive  2010 target_demo                               NA
#> 10 automotive  2011 target_demo                               NA
#> # … with 123 more rows
#> 
#> $target_NA
#> # A tibble: 133 x 4
#>    sector      year emission_factor_metric emission_factor_value
#>    <chr>      <dbl> <chr>                                  <dbl>
#>  1 automotive  2002 target_NA                                 NA
#>  2 automotive  2003 target_NA                                 NA
#>  3 automotive  2004 target_NA                                 NA
#>  4 automotive  2005 target_NA                                 NA
#>  5 automotive  2006 target_NA                                 NA
#>  6 automotive  2007 target_NA                                 NA
#>  7 automotive  2008 target_NA                                 NA
#>  8 automotive  2009 target_NA                                 NA
#>  9 automotive  2010 target_NA                                 NA
#> 10 automotive  2011 target_NA                                 NA
#> # … with 123 more rows
#> 

# Calculate company-level targets
out <- target_sda(
  matched, ald2,
  co2_intensity_scenario = co2_scenario,
  by_company = TRUE
)